Исправляет ошибки с компонентами события
